Reviewed and checked the export of the creation of the entire gushim.topojson include the file topojson file

+++ gushim/program.py
@@ -1,3 +1,4 @@
+import datetime
 import logging
 import logging.config
 import pandas as pd
@@ -5,15 +6,17 @@ import geopandas as gp
 import numpy as np
 from shapely.geometry import Point, Polygon, LineString
 import time
+from io import open as io_open
 import os
 import sys
 import yaml
+from git import Repo
 
 sys.path.insert(0, 'opentaba-gushim-prj')
 import gushim.mapi_service as mapi
 import gushim.compress as compress
 import gushim.geo_utils as geo_utils
-import gushim.utilities as util
+# import gushim.utilities as util
 
 gushim_url ='https://data.gov.il/dataset/3ef36ec6-f0e3-447b-bc9d-9ab4b3cee783/resource/f7a10a68-fba5-430b-ac11-970611904034/download/subgushall-nodes.zip'
 
@@ -21,14 +24,20 @@ gushim_url ='https://data.gov.il/dataset/3ef36ec6-f0e3-447b-bc9d-9ab4b3cee783/re
 END_NODE_FILE = 'nodes.csv'
 END_ATTRIBUTE_FILE = 'attributes01.csv'
 YESHUV_MASK_FILE = r'support_data/Yeshuvim2015.csv'
-MIN_POPULATION = 2000
-SAVE_GUSHIM_SHAPEFILE = False  # Save the gushim to shapefile
-EXPORT_TO_GEOJSON = False
-EXPORT_TO_TOPOJSON = True  # Save to TopoJSON
-EXPORT_ALL_GUSHIM = False
 WORKSPACE_FOLDER = 'workspace'
 EXPORT_FOLDER = 'export'
+MIN_POPULATION = 2000
 LOAD_SHAPEFILE = True
+SHAPEFILE_PATH = r'./export/Gushim20170222_110536.shp'
+SAVE_GUSHIM_SHAPEFILE = False  # Save the gushim to shapefile
+EXPORT_TO_GEOJSON = False
+EXPORT_TO_TOPOJSON = False  # Save to TopoJSON
+EXPORT_ALL_GUSHIM = True
+REPO_DIR = '../'
+NEW_BRANCH = False
+PUSH_TO_GITHUB = False
+
+
 
 
 def setup_logging(default_path='logging.yaml', default_level=logging.INFO, env_key='LOG_CFG'):
@@ -55,13 +64,16 @@ def get_mapi_uncompress_file(folder_path, mapi_format):
     :param mapi_format: The file format of the gushim unzipped file - currently csv
     :return: return filename
     """
-    for (dirpath, dirnames, filenames) in os.walk(folder_path):
-        csv_files = [os.path.join(dirpath, fi) for fi in filenames if fi.endswith(mapi_format)][0]
-
+    for (dir_path, dir_names, file_names) in os.walk(folder_path):
+        csv_files = [os.path.join(dir_path, fi) for fi in file_names if fi.endswith(mapi_format)][0]
     return csv_files
 
 
 def get_or_create_folder(folder_name):
+    """
+    :param folder_name:
+    :return:
+    """
     base_folder = os.path.abspath(os.path.dirname(__file__))
     if not folder_name:
         folder_name = 'workspace'
@@ -76,6 +88,8 @@ def get_or_create_folder(folder_name):
 
 def main():
     logger = setup_logging(default_level=logging.DEBUG)
+    datetime_str = datetime.datetime.utcnow().strftime('%Y%m%d_%H%M')
+
     logger.info("Program started")
 
     workspace_folder = get_or_create_folder(WORKSPACE_FOLDER)
@@ -83,6 +97,10 @@ def main():
     logger.debug('The workspace folder is: {0}'.format(workspace_folder))
 
     base_folder = os.path.abspath(os.path.dirname(__file__))
+    repo = Repo(REPO_DIR)
+    if NEW_BRANCH:
+        origin = repo.create_remote(datetime_str, repo.remotes.origin.url)
+        repo.create_head(datetime_str, origin.refs.master)
 
     if not LOAD_SHAPEFILE:
         # Load Yeshuvim mask data into pandas
@@ -134,45 +152,56 @@ def main():
         df_polygon_att_wgs = df_polygon_att.to_crs({'init': 'epsg:4326'}).copy()
         logger.info('Successfully projected file GeoDataframe to WGS84')
     else:
-        shapefile_path = r'C:\Users\sephi\github\opentaba_gushim\opentaba_gushim_prj\gushim\export\Gushim20170209_173107.shp'
-        df_polygon_att_wgs = gp.read_file(shapefile_path)
-        logger.info('Shapefile {} loaded into GeoPandas'.format(shapefile_path ))
+        df_polygon_att_wgs = gp.read_file(SHAPEFILE_PATH)
+        logger.info('Shapefile {} loaded into GeoPandas'.format(SHAPEFILE_PATH))
+
+    if SAVE_GUSHIM_SHAPEFILE:
+        # timestr = time.strftime("%Y%m%d_%H%M%S")
+        export_file = os.path.join(base_folder, export_folder, 'Gushim{0}.shp'.format(datetime_str))
+        df_polygon_att_wgs.to_file(export_file, encoding="utf-8")
+        logger.info('Successfully saved shapefile {0}'.format(export_file))
 
     # Get list of localities
     localities = df_polygon_att_wgs.groupby(['EngName']).size().index.tolist()
     # Prepare dataframe with relevant topojson column
     df_polygon_att_wgs['Name'] = df_polygon_att_wgs['GUSH_NUM']  # df_local.rename(columns={'GUSH_NUM': 'name'}, inplace=True)
 
-    # export each locality as geojson
-    for local in localities:
-        df_local = df_polygon_att_wgs[(df_polygon_att_wgs.EngName == local) & (df_polygon_att_wgs.Pop2015 > MIN_POPULATION)]
-        if not df_local.empty:
-            file_name_string = "".join(x for x in local if x.isalnum()).encode('utf-8').upper()  #remove unsafe characters
-            try:
-                export_file = os.path.join(base_folder, export_folder, '{0}.geojson'.format(file_name_string))
-                if EXPORT_TO_GEOJSON:
-                    local_geojson = df_local.to_json()
-                    with open(export_file, 'w') as f:
-                        f.write(local_geojson)
-                if EXPORT_TO_TOPOJSON:
-                    export_temp_geojson_file = os.path.join(base_folder, export_folder, 'temp_json4{0}.geojson'.format(file_name_string))
-                    local_geojson = df_local[['Name', 'geometry']].to_json()  # local_topojson = df_local.to_json()
-                    # if not os.path.exists(export_file) or not os.path.isdir(export_file):
-                    with open(export_temp_geojson_file, 'w') as f:
-                        f.write(local_geojson)
-                    # filename, _ = os.path.splitext(export_temp_geojson_file)
-                    topojson_file = os.path.join(base_folder, export_folder, '{0}.topojson'.format(file_name_string))
-                    geo_utils.geoson_to_topojson(export_temp_geojson_file, topojson_file)
-                    try:
-                        os.remove(export_temp_geojson_file)
-                        logger.debug('Successfully deleted temp_geojson file: {0}'.format(export_temp_geojson_file))
-                    except OSError, e:
-                        print ("Error: {} - {}.".format(e.message, e.strerror))
 
-            except OSError, e:
-                print ("Error: {} - {}.".format(e.message, e.strerror))
-                logger.warning('failed to save geojson for: {0}'.format(local))
-    logger.debug('Finished to saved files to GeoJSON')
+    # export each locality as geojson
+    if EXPORT_TO_GEOJSON or EXPORT_TO_TOPOJSON:
+        for local in localities:
+            df_local = df_polygon_att_wgs[(df_polygon_att_wgs.EngName == local) & (df_polygon_att_wgs.Pop2015 > MIN_POPULATION)]
+            if not df_local.empty:
+                file_name_string = "".join(x for x in local if x.isalnum()).encode('utf-8').lower()  #remove unsafe characters
+                try:
+                    export_file = os.path.join(base_folder, export_folder, '{0}.geojson'.format(file_name_string))
+                    if EXPORT_TO_GEOJSON:
+                        local_geojson = df_local.to_json()
+                        with io_open(export_file, 'w', encoding='utf-8') as f:
+                            f.write(unicode(local_geojson))
+                        # with io_open(export_file, 'w', encoding='utf-8') as f:
+                        #     df_local.to_json(f, force_ascii=False)
+                    if EXPORT_TO_TOPOJSON:
+                        export_temp_geojson_file = os.path.join(base_folder, export_folder, 'temp_json4{0}.geojson'.format(file_name_string))
+                        local_geojson = df_local[['Name', 'geometry']].to_json()  # local_topojson = df_local.to_json()
+                        # if not os.path.exists(export_file) or not os.path.isdir(export_file):
+                        with open(export_temp_geojson_file, 'w') as f:
+                            f.write(local_geojson)
+                        # filename, _ = os.path.splitext(export_temp_geojson_file)
+                        topojson_file = os.path.join(base_folder, export_folder, '{0}.topojson'.format(file_name_string))
+                        geo_utils.geoson_to_topojson(export_temp_geojson_file, topojson_file)
+                        try:
+                            os.remove(export_temp_geojson_file)
+                            logger.debug('Successfully deleted temp_geojson file: {0}'.format(export_temp_geojson_file))
+                        except OSError, e:
+                            print ("Error: {} - {}.".format(e.message, e.strerror))
+
+                except OSError, e:
+                    print ("Error: {} - {}.".format(e.message, e.strerror))
+
+
+                    logger.warning('failed to save geojson for: {0}'.format(local))
+        logger.debug('Finished to saved files to GeoJSON')
 
     if EXPORT_ALL_GUSHIM:
         all_gushim_file = os.path.join(base_folder, export_folder, '{0}.geojson'.format('ISRAEL_GUSHIM'))
@@ -185,11 +214,6 @@ def main():
 
     # TODO convert to function taking dfpg and crs
 
-    if SAVE_GUSHIM_SHAPEFILE:
-        timestr = time.strftime("%Y%m%d_%H%M%S")
-        export_file = os.path.join(base_folder, export_folder, 'Gushim{0}.shp'.format(timestr))
-        df_polygon_att_wgs.to_file(export_file)
-        logger.info('Successfully saved shapefile {0}'.format(export_file))
 
 
     logger.info('Successfully  created  Polygon in geopandas')
@@ -205,6 +229,16 @@ def main():
     logger.debug('Number of Gushim per Localitis is {0}'.format(df_polygon_att_wgs.groupby(['HebName'])['GUSH_NUM'].size()))
 
     # TODO Push to GIT
+    if PUSH_TO_GITHUB:
+        repo = Repo(REPO_DIR)
+        topojson_files = [i for i in repo.untracked_files if i.endswith('topojson')]
+        repo.index.add(topojson_files)
+        repo.index.commit('Files created on {0}'.format(datetime_str))
+        remote_origin = repo.remote('origin')
+        remote_origin.push(repo.active_branch.name)
+        logger.info('Files where pushed to GitHub repo')
+        logger.debug('The following files were pushed to the repo {0}'.format(topojson_files))
+
 
     logger.info("Program completed")
